fix(subscribeOn): allow Infinity as valid delay
Removes use of `isNumeric` that excludes `Infinity` as a valid number
- Moves SubscribeOnObservable to the only location it is used.
- Removes redundant tests of internals
