+++ spec/operators/subscribeOn-spec.ts
@@ -79,4 +79,14 @@ describe('subscribeOn operator', () => {
     expectObservable(result, unsub).toBe(expected);
     expectSubscriptions(e1.subscriptions).toBe(sub);
   });
+
+  it('should properly support a delayTime of Infinity', () => {
+    const e1 =   hot('--a--b--|');
+    const expected = '---------';
+    // TODO: Support lack of subscription in `expectSubscriptions`
+    // const sub =      '         ';
+
+    expectObservable(e1.pipe(subscribeOn(rxTestScheduler, Infinity))).toBe(expected);
+    // expectSubscriptions(e1.subscriptions).toBe(sub);
+  });
 });
