+++ src/internal/util/isScheduler.ts
@@ -1,5 +1,5 @@
 import { SchedulerLike } from '../types';
 
 export function isScheduler(value: any): value is SchedulerLike {
-  return value && typeof (<any>value).schedule === 'function';
+  return value && typeof value.schedule === 'function';
 }
